Fix major errors
