+++ src/security/index.ts
@@ -1,6 +1,3 @@
-export * from "./baseVisibilityContextCommandProvider";
-export * from "./popupSecurityModelEditorProvider";
-export * from "./roleBasePopupSecurityModelEditorProvider";
 export * from "./security.design.module";
 export * from "./security.module";
 export * from "./securityModelEditor";
