--- src/security/roleBasePopupSecurityModelEditorProvider.ts
@@ -1,8 +1,0 @@
-import { ViewManager } from "@paperbits/common/ui";
-import { PopupSecurityModelEditorProvider } from "./popupSecurityModelEditorProvider";
-
-export class RoleBasePopupSecurityModelEditorProvider extends PopupSecurityModelEditorProvider {
-    constructor(viewManager: ViewManager) {
-        super(viewManager, "role-based-security-model-editor");
-    }
-}
