+++ src/security/security.design.module.ts
@@ -2,7 +2,7 @@ import { IInjector, IInjectorModule } from "@paperbits/common/injection";
 import { AccessContextToolButton } from "../workshops/accessContextToolbutton";
 import { RoleBasedSecurityModelEditor } from "../workshops/roles/ko/roleBasedSecurityModelEditor";
 import { SecurityModule } from "./security.module";
-import { RoleBasePopupSecurityModelEditorProvider } from "./roleBasePopupSecurityModelEditorProvider";
+import { RoleBaseSecurityModelEditorProvider } from "./roleBaseSecurityModelEditorProvider";
 
 export class SecurityDesignModule implements IInjectorModule {
     register(injector: IInjector): void {
@@ -10,6 +10,6 @@ export class SecurityDesignModule implements IInjectorModule {
 
         injector.bindToCollection("trayCommands", AccessContextToolButton);
         injector.bind("securityModelEditor", RoleBasedSecurityModelEditor);
-        injector.bind("visibilityCommandProvider", RoleBasePopupSecurityModelEditorProvider);
+        injector.bind("visibilityCommandProvider", RoleBaseSecurityModelEditorProvider);
     }
 }
